[IMP] l10n_ar_sale: price unit precision digits on sale report.
Description of the issue/feature this PR addresses:
If "Product Price" Decimal Accuracy has more than 2 digits, then the unit price on the printed report has only 2 digits when Argentinean sale report is printed and this causes that quantity * price on each invoice line does not result the amount shown on the "Amount" column. This issue was fixed for invoices but not for sale orders: https://github.com/odoo/odoo/pull/171131 .

Steps to reproduce:
1. Log ing with admin user on adhoc runbot 18 instance, activate developer mode and install l10n_ar_sale module.
2. Take position on "Muebleria ARG" company, go to menu "Setting / Technical / Database Structure / Decimal Accuracy" and set 4 digits for "Product Price".
3. Create a sale order to the partner "Consumidor Final Anónimo" (if the partner is "Consumidor Final Anónimo" then taxes are not discriminated on the printed sale report --> document type = "FACTURAS B"). The sale order must have an invoice line with product price with 4 digits, for example: quantity 3000, price 65.3057 and tax "IVA 21%".
4. Confirm and print the sale order.
The result of quantity * price is not the amount shown on the printed invoice (3000 * 79.02 != 237059.69)

Current behavior before PR:
If "Product Price" Decimal Accuracy has more than 2 digits, then the unit price on the printed report has only 2 digits when Argentinean sale report is printed and this causes that quantity * price on each invoice line does not result the amount shown on the amount column.

Desired behavior after PR is merged:
If "Product Price" Decimal Accuracy has more than 2 digits, then the unit price on the printed report has the same quantity of digits set on "Product Price" Decimal Accuracy when Argentinean sale report is printed.

Task Adhoc side: 55629

closes ingadhoc/argentina-sale#255

Signed-off-by: Julia Elizondo <jue@adhoc.com.ar>